Don't require java package naming conventions
SBT does not enforce the java package naming convention for source
files (e.g. the source file Bar.scala for package com.meraki.foo would
be found in $TOPDIR/src/main/scala/com/meraki/foo/Bar.scala). This,
however, is where the jacoco sbt plugin requries source file to be if it
is to include Bar.scala.html in the html output. To get around this
restriction, I now allow jacoco to look in the root source directory as
well as the package directory for the desired file.